Formating, moved to new directory, rename class, ternary if, secure random Formátování, přesunuto do nové složky, přejmenování class, ternární podmínka, secure random místo random
